Fix the build

+++ src/arduinosinled.cpp
@@ -11,7 +11,7 @@ SinLed::SinLed(const uint8_t& pin, const float& step) :
 void SinLed::initialize(const uint8_t& ledstartupcount, const int& delayms) {
   pinMode(pin_, OUTPUT);
   uint16_t step = 0xff / ledstartupcount;
-  uint8_t led = step;
+  uint8_t led = static_cast<uint8_t>(step);
   for (uint8_t n = 0; n < ledstartupcount; n++) {
     analogWrite(pin_, led);
     delay(delayms);
@@ -31,7 +31,7 @@ void SinLed::setstep(const float& step) {
 
 uint8_t SinLed::cycle(const uint8_t& max) {
   uint8_t value = outputsin(max);
-  at_ = at_ < FDS_LED_SIN_MAXIMUM_AT ? at_ + step_ : 0.0;
+  at_ = at_ < FDS_LED_SIN_MAXIMUM_AT ? at_ + step_ : 0.0F;
   return value;
 }
 
@@ -53,7 +53,7 @@ void SinLed::fullcycle(const uint8_t& count, const unsigned long& delayms, const
 }
 
 uint8_t SinLed::outputsin(const uint8_t& max) {
-  float half = static_cast<double>(max) / 2.0;
+  float half = static_cast<float>(max) / 2.0F;
   uint8_t value = static_cast<uint8_t>(sin(at_) * half + half);
   analogWrite(pin_, value);
   return value;
